Extract `NoiseSessionManager` into a separate file
